requirements.txt
numpy<1.24
transforms3d
gym==0.25.2
sapien==2.2.1
open3d
imageio
nlopt
chumpy
opencv-python
tyro
pyyaml
trimesh
anytree
lxml
pycollada
stable-baselines3==1.1.0
loguru
lockfile
gdown
yacs
glfw
scikit-learn
pillow
pyquaternion
scipy
tabulate
tqdm
moviepy
natsort
absl-py
easydict
multimethod
termcolor
shortuuid
fire
hydra-core
hydra-ray-launcher
hydra-submitit-launcher
wandb
tensorboard
grpcio-tools
einops==0.4.1
diffusers==0.11.1
numba==0.56.4
dill==0.3.5.1
zarr
setuptools==59.5.0
Cython==0.29.35
patchelf==0.17.2.0
